refactor(map): separar clase Tile en archivo dedicado
- Tile fue movida a roguelike_project/map/tile.py siguiendo SRP.
- map_loader.py ahora solo se encarga de la carga de datos del mapa.
- Se mejora la organización del código y se prepara para futuras extensiones (tiles animados, interactivos).
