+++ roguelike_project/map/tile.py
@@ -1,0 +1,25 @@
+import pygame
+
+from roguelike_project.config import DEBUG
+from roguelike_project.config import TILE_SIZE
+
+class Tile:
+    def __init__(self, x, y, tile_type, sprite):
+        self.x = x
+        self.y = y
+        self.tile_type = tile_type
+        self.sprite = sprite
+        self.solid = tile_type == "#"
+        self.rect = pygame.Rect(x, y, TILE_SIZE, TILE_SIZE)
+
+    def render(self, screen, camera):
+        if camera is None:
+            screen.blit(self.sprite, (self.x, self.y))
+            return pygame.Rect(self.x, self.y, TILE_SIZE, TILE_SIZE)
+
+        scaled_sprite = pygame.transform.scale(self.sprite, camera.scale((self.sprite.get_width(), self.sprite.get_height())))
+        screen.blit(scaled_sprite, camera.apply((self.x, self.y)))
+
+        if self.solid and DEBUG:
+            scaled_rect = pygame.Rect(camera.apply(self.rect.topleft), camera.scale(self.rect.size))
+            pygame.draw.rect(screen, (255, 255, 0), scaled_rect, 1)
